added all of the stuff from liban's form events to trevor's form events file which is located in the events folder
